Minor refactoring. Added unit tests to achieve better unit test coverage.
