+++ test/Dapr.Jobs.Test/Models/DaprJobScheduleTests.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using Dapr.Jobs.Models;
 using Xunit;
 
 namespace Dapr.Jobs.Test.Models;
